scripts/requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
Pillow==10.3.0
